pyproject.toml
[project]
name = "my-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "python-telegram-bot>=22.5",
]
